backend/requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
langchain-core>=0.1.10
langchain-chroma>=0.1.0
openai>=1.12.0
pydantic>=2.5.2
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy>=2.0.25
pymysql>=1.1.0
alembic>=1.13.1
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
bcrypt>=4.1.2
